Step2 공유

+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,5 +1,10 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberService;
+import nextstep.support.AuthorizationException;
+import nextstep.ui.AuthenticationPrincipal;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,13 +16,19 @@ import java.util.List;
 public class ReservationController {
 
     public final ReservationService reservationService;
+    public final MemberService memberService;
 
-    public ReservationController(ReservationService reservationService) {
+    public ReservationController(ReservationService reservationService, MemberService memberService) {
         this.reservationService = reservationService;
+        this.memberService = memberService;
     }
 
     @PostMapping
-    public ResponseEntity createReservation(@RequestBody ReservationRequest reservationRequest) {
+    public ResponseEntity createReservation(@AuthenticationPrincipal String token, @RequestBody ReservationRequest reservationRequest) {
+        if (token == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        Member member = memberService.findByToken(token);
         Long id = reservationService.create(reservationRequest);
         return ResponseEntity.created(URI.create("/reservations/" + id)).build();
     }
@@ -29,7 +40,18 @@ public class ReservationController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable Long id) {
+    public ResponseEntity deleteReservation(@AuthenticationPrincipal String token, @PathVariable Long id) {
+        if (token == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        // 권한 체크(본인의 예약건인지 확인)
+        Member member = memberService.findByToken(token);
+        Reservation reservation = reservationService.findById(id);
+        if (!reservation.getName().equals(member.getName())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         reservationService.deleteById(id);
 
         return ResponseEntity.noContent().build();
